Moved functionality for moving pieces to the MovePieceInteractor, changed types of input and output boundaries to match expected types
