+++ src/main/java/interfaceAdapters/legalMoves/LegalMovesState.java
@@ -4,5 +4,5 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class LegalMovesState {
-    public List<String> legalMoves = new ArrayList<>();
+    public List<Integer> legalMoves = new ArrayList<>();
 }
